pyproject.toml
[tool.poetry]
name = "strava"
version = "0.1.0"
description = ""
authors = ["voynow <voynow99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
stravalib = "^v1.7"
supabase = "*"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
